+++ trainingApp/models.py
@@ -5,23 +5,6 @@ from django.utils import timezone
 from django.contrib.auth.models import User
 
 
-class Trainee(models.Model):
-    #Enumeracion para el nivel educativo del user
-    class EducationalLevel(models.TextChoices):
-        Primary_School = 'Primaria'
-        High_School = 'Secundaria'
-        Collage = 'Universidad'
-        Graduate = 'Postgrado'  
-  
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    age =  models.IntegerField(default=18)
-    educationalLevel = models.CharField(
-        max_length=20,
-        choices=EducationalLevel.choices,
-        default=EducationalLevel.High_School
-    )
-    occupation = models.CharField(max_length=50)
-
 class Training(models.Model):
     #Enumeracion para el tipo de entrenamiento
     class TrainingType(models.TextChoices):
