+++ userApp/templates/userApp/base.html
@@ -1,3 +1,4 @@
+<!-- Se cargan los archivos estaticos como hojas de estilos e imagenes -->
 {% load static %}
 
 <!DOCTYPE html>
@@ -8,7 +9,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js"></script>
-    <link rel="stylesheet" href="{% static 'userApp/style/style.css' %}">
+    <link rel="stylesheet" href="{% static 'userApp/style/style.css' %}">    <!-- Se utiliza un archivo estatico con la sintaxis  -->
     <title>List of trainings</title>
   </head>
   <body>
